# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Future API OpenApi Documentation",
        "description": "Future API OpenApi description",
        "license": {
            "name": "GNU General Public License v3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Sandbox server (uses test data)"
        }
    ],
    "paths": {
        "/notebook": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Get list of notebooks",
                "description": "Returns list of notebooks",
                "operationId": "getNotebooksList",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Requested-With"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Desired number of entries per page, between 3 and 100",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 3,
                            "example": "25"
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "description": "Cursor for cursor pagination",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "eyJub3RlYm9va3MuaWQiOiI5NmUxY2I4YS1jMzVmLTRkZjAtOWI3Ni02MjlmNDNjZTFkNDUiLCJfcG9pbnRzVG9OZXh0SXRlbXMiOnRydWV9"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetNotebooks200"
                                }
                            }
                        }
                    },
                    "405": {
                        "$ref": "#/components/responses/405"
                    },
                    "429": {
                        "$ref": "#/components/responses/429"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Add single notebook",
                "description": "Adds single notebook.",
                "operationId": "addNotebookEntry",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Requested-With"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/addNotebookRequest"
                },
                "responses": {
                    "201": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddNotebook201"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "405": {
                        "$ref": "#/components/responses/405"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "429": {
                        "$ref": "#/components/responses/429"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Add single user",
                "description": "Adds single user.",
                "operationId": "addUserEntry",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Requested-With"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/addUserRequest"
                },
                "responses": {
                    "201": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddUser201"
                                }
                            }
                        }
                    },
                    "405": {
                        "$ref": "#/components/responses/405"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "429": {
                        "$ref": "#/components/responses/429"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login user",
                "description": "Logs user in.",
                "operationId": "loginUser",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Requested-With"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/loginUserRequest"
                },
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginUser200"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "405": {
                        "$ref": "#/components/responses/405"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "429": {
                        "$ref": "#/components/responses/429"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Logout user",
                "description": "Logs user out.",
                "operationId": "logoutUser",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Requested-With"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LogoutUser200"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "405": {
                        "$ref": "#/components/responses/405"
                    },
                    "429": {
                        "$ref": "#/components/responses/429"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Name": {
                "description": "Person's name.",
                "type": "string",
                "maxLength": 255,
                "example": "John Doe"
            },
...
            "AddNotebook201": {
                "description": "Notebook object has been added successfully",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/EntryUuid"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/SrvUpdatedAt"
                    },
                    "message": {
                        "$ref": "#/components/schemas/SrvMessage"
                    }
                },
                "type": "object"
            },
...
            "GetNotebooks200": {
                "description": "Notebook object has been returned successfully",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "$ref": "#/components/schemas/EntryUuid"
                                },
                                "creator_uuid": {
                                    "$ref": "#/components/schemas/EntryUuid"
                                },
                                "name": {
                                    "$ref": "#/components/schemas/Name"
                                },
                                "company": {
                                    "$ref": "#/components/schemas/Company"
                                },
                                "phone": {
                                    "$ref": "#/components/schemas/Phone"
                                },
                                "email": {
                                    "$ref": "#/components/schemas/Email"
                                },
                                "birthday": {
                                    "$ref": "#/components/schemas/Birthday"
                                },
                                "created_at": {
                                    "$ref": "#/components/schemas/SrvCreatedAt"
                                },
                                "updated_at": {
                                    "$ref": "#/components/schemas/SrvUpdatedAt"
                                },
                                "picture_url": {
                                    "$ref": "#/components/schemas/Url"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "path": {
                        "$ref": "#/components/schemas/GetNotebooksPath"
                    },
                    "per_page": {
                        "$ref": "#/components/schemas/GetNotebooksPerPage"
                    },
                    "next_cursor": {
                        "$ref": "#/components/schemas/GetNotebooksNCursor"
                    },
                    "next_page_url": {
                        "$ref": "#/components/schemas/GetNotebooksNPage"
                    },
                    "prev_cursor": {
                        "$ref": "#/components/schemas/GetNotebooksPCursor"
                    },
                    "prev_page_url": {
                        "$ref": "#/components/schemas/GetNotebooksPPage"
                    }
                },
                "type": "object"
            },
...
            "AddUser201": {
                "description": "User object has been added successfully",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/EntryUuid"
                    },
                    "access_token": {
                        "$ref": "#/components/schemas/AccessToken"
                    },
                    "token_type": {
                        "$ref": "#/components/schemas/AccessTokenType"
                    },
                    "message": {
                        "$ref": "#/components/schemas/SrvMessage"
                    }
                },
                "type": "object"
            },
...
            "LogoutUser200": {
                "description": "User has logged out successfully",
                "properties": {
                    "message": {
                        "$ref": "#/components/schemas/SrvMessage"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "401": {
                "description": "User is not authenticated.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "$ref": "#/components/schemas/SrvMessage"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "405": {
                "description": "Method not allowed for this endpoint.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "$ref": "#/components/schemas/SrvMessage"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "422": {
                "description": "Error in values sent to server. Check server reply for more info.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "$ref": "#/components/schemas/SrvMessage"
                                },
                                "errors": {
                                    "properties": {
                                        "fieldName": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Human-readable explanation, what's wrong with provided field value."
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "429": {
                "description": "Too Many Requests. Please, respect x-ratelimit-limit responce header value.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "$ref": "#/components/schemas/SrvMessage"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "500": {
                "description": "Internal Server Error. We'll try to fix it ASAP.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "$ref": "#/components/schemas/SrvMessage"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "X-Requested-With": {
                "name": "X-Requested-With",
                "in": "header",
                "description": "Should be equal to XMLHttpRequest",
                "required": true,
                "schema": {
                    "type": "string",
                    "enum": [
                        "XMLHttpRequest"
                    ]
                }
            },
...
        },
        "requestBodies": {
            "addNotebookRequest": {
                "description": "Notebook object to add to server",
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "required": [
                                "name",
                                "phone",
                                "email"
                            ],
                            "properties": {
                                "id": {
                                    "$ref": "#/components/schemas/EntryUuid"
                                },
                                "name": {
                                    "$ref": "#/components/schemas/Name"
                                },
                                "company": {
                                    "$ref": "#/components/schemas/Company"
                                },
                                "phone": {
                                    "$ref": "#/components/schemas/Phone"
                                },
                                "email": {
                                    "$ref": "#/components/schemas/Email"
                                },
                                "birthday": {
                                    "$ref": "#/components/schemas/Birthday"
                                },
                                "picture": {
                                    "$ref": "#/components/schemas/Picture"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "addUserRequest": {
                "description": "User object to add to server",
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "required": [
                                "name",
                                "email",
                                "password"
                            ],
                            "properties": {
                                "name": {
                                    "$ref": "#/components/schemas/Name"
                                },
                                "email": {
                                    "$ref": "#/components/schemas/Email"
                                },
                                "password": {
                                    "$ref": "#/components/schemas/Password"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "loginUserRequest": {
                "description": "User credentials to log in to server",
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "required": [
                                "email",
                                "password"
                            ],
                            "properties": {
                                "email": {
                                    "$ref": "#/components/schemas/Email"
                                },
                                "password": {
                                    "$ref": "#/components/schemas/Password"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}